Add RADOLAN method and example notebook with OpenMRG data (#6)
* docs/notebooks/data/openMRG_create.ipynb:
 - creates 5min aggregated CML, rain gauge and radar data for the openmerge dataset

src/mergeplg/merge.py:
 - first implementation of merging

src/mergeplg/interpolator.py and src/mergeplg/idw.py:
 - IDW interpolator functions from python

* add RADOLAN functions copy-paste

* remove verion caps

* remove stuff from Erlend's PR that I do not need. I only need the example data.

Could have done that in a better way, but try to move forward quickly now.

* dummy commit to trigger CI

* a lot of linting fixed in adjust.py

* linting fixed in check_dats_struct.py

* temporarily add pycomlink as submodule

I do this so that I can run the copy-pasted code with the OpenMRG example data here, visually check the results and then adjust the test that rely on DWD data to use the OpenMRG data. Only then I will change the code to use polirgain instaead of wradlib and pycomlink code.

* Set max Python version to 4.0.0 because nbmake has this version capping

* linting fixed for processing.py

doc strings still have to be filled with info

* minor things

* add IDW code

* add numba as dependency and remove some version caping in toml file

* add pycomlink submodule as dev dependency in poetry

Note sure how that will affect a potential later install via pip. But we have to switch to a pip-installable pycomlink version later on anyway.

* enable fetching submodules in CI config

* exclude submodules in readthedocs

* use grid_intersection code from poligrain and remove some pycomlink submodule stuff

* disable pylint for processing.py

* pylint error, move import

* added working, but WIP, RADOLAN example notebook

 Note that the projected coordinates are in meters. Hence, the default of `max_distance=60` does not make sense. In the example I use 30e3 meters.

* added validation with city gauges to notebook

Clear improvement of radar rainfall fields with RADOLAN compared to raw radar. Similar to what Erlend shows in his notebook, on which my noteobook is based.

* dev in notebook for adjusting tests...

* added radolan module files to __init__

* added test for label_relevant_audit_interim_in_gageset

These are mofified test from old code. I checked locally that all these test fail when adjusting variables. Hence, they are fine.

* added test for get_grid_rainfall_at_points

This is a variation of a test from old code. I checked locally that it fails when forced to do so. Hence, this test is fine.

* add test for interpolate_station_values

variation of old test for old code. Checked that if fails if forced to. Good.

* added test for bogra_like_smoothing

copy-paste of old test work 1:1 with the code, and reduced test data, here

* update notebook with dev part of test, which need to be removed later

maybe already added in commits before...

* clean up RADOLAN OpenMRG example

not final, but a lot cleaner than before

* remove  PD011 because this is now disable in toml file

* add test data (which I forgot to add before...) and adjusted function to load test data from correct file names

* addjusted test data path

* fix problem with sorting, should now work using `stable` sorting

* undo dummy change

* add test for raise condition of label_relevant_audit_interim

* add test for interpolation with only nan values

* rename radolan adjust test file

* add first test for radolan processing module

* Added some plausibilisation of RW intermediate results to RADOLAN
details notebook to make sure that the results I use as reference in the
tests make sense.

* added tests for rh_to_rw

these test did need a lot of adjustment because we do not use the station data from neighboring countries here and unfortunately in my old codebase I did chose a region at the border to test correctness of results. Results are now different in these regions due to missing adjustment stations. Hence, I did chose new regions for checking results. Results where validated visually in the notebook and looked good to me.

* ignore deprecation warning for specific xarray call via pyproject.toml config of pytest

This should be removed later

* add function to load and transform OpenMRG data, also added test

* added transform function for CML data from Dataset to DataFrame for old RADOLAN code

also added tests

* removed RADOLAN details notebook for now because it is so messy and fixing linting will  take too long.

Will add it back later in a cleaner state.

* call pytest directly in CI instead of python -m pytest

* Revert "call pytest directly in CI instead of python -m pytest"

This reverts commit fb3e018ea23f59cc9d2a1f9c5c546078491c067c.

* add init.py to test dir to allow relative imports there

* using a copy of the OpenMRG example data in the src dir

I have found no other way to make the load with the existing io function work on CI because the data did live outside the src dir

* cleand up RADOLAN OpenMRG notebook and add it to doc

* typo...

* fix syntax highlighting in readthedocs

* final docstring in rh_to_rw

* add test for IDW

* run example notebook again

---------

Co-authored-by: eoydvin <erlend.oydvin@gmail.com>